fix: no middle token line detection
This commit resolves a bug where a four-in-a-row connection would only
be detected if the forming token was dropped at either end of the line,
but not if it was dropped somewhere in the middle of the line.

The logic now resorts to evaluating every possible line for every cell
on the board each time a player drops a token into the board. This is
far less efficient and elegant than the previous logic which checked
for connections based only on the most recently dropped token.

That said, one advantage to this new design is that it more easily
allows for connection checking outside of a specific player's turn,
leading to more possible method extractions in the Game class.

In addition to this bug fix, the script methods responsible for
coordinating the gameplay on the command line have been detached from
Game into their own CLI class.